integrated beam module with instrument package
